🛠️ Refactor: Toasts, SharedChatFields, and Intl - oh my! (#81)
* SharedChatFields, move github req to network, intl for ban user, constants/toasts

* move constants/messages into content/

* add toast for logout
