+++ client/content/messages/components/CreateChatModal.js
@@ -1,20 +1,10 @@
 import { defineMessages } from "react-intl";
 
-import locales from "../../../content/locale";
+import locales from "../../locale";
 
 // Note: these are also used by EditChatModal
 
 export const messages = defineMessages({
-  name: {
-    id: "name",
-    description: locales.en.name,
-    defaultMessage: locales.en.name,
-  },
-  description: {
-    id: "description",
-    description: locales.en.description,
-    defaultMessage: locales.en.description,
-  },
   link: {
     id: "link",
     description: locales.en.link,
@@ -74,11 +64,6 @@ export const messages = defineMessages({
     id: "community",
     description: locales.en.community,
     defaultMessage: locales.en.community,
-  },
-  gcNameTip: {
-    id: "gc-name-tip",
-    description: locales.en["gc-name-tip"],
-    defaultMessage: locales.en["gc-name-tip"],
   },
   submitGroupChat: {
     id: "submit-group-chat",
